FaceQuadrature: add compilation unit
